+++ js/main.js
@@ -1,6 +1,6 @@
 import { photosDataset } from './data.js';
 import { renderPhotos } from './thumbnails.js';
 import './big-picture-renderer.js';
-import './upload-image-form-renderer.js';
+import './form-renderer.js';
 
 renderPhotos(photosDataset);
